+++ Source/Lib/ASM_AVX2/temporal_filtering_avx2.c
@@ -199,7 +199,7 @@ static void svt_av1_apply_zz_based_temporal_filter_planewise_medium_partial_avx2
             __m128i frame2_array      = _mm_loadl_epi64((__m128i*)(y_pre + k));
             frame2_array              = _mm_cvtepu8_epi16(frame2_array);
             __m256i frame2_array_u32  = _mm256_cvtepi16_epi32(frame2_array);
-            frame2_array_u32          = _mm256_mullo_epi32(frame2_array_u32,
+            frame2_array_u32 = _mm256_mullo_epi32(frame2_array_u32,
                                                   adjusted_weight_int32[subblock_idx_h + (j >= block_width / 2)]);
 
             accumulator_array = _mm256_add_epi32(accumulator_array, frame2_array_u32);
@@ -328,7 +328,7 @@ static void svt_av1_apply_temporal_filter_planewise_medium_partial_avx2(
                                        block_error_fp8[subblock_idx]) /
             (TF_WINDOW_BLOCK_BALANCE_WEIGHT + 1);
 
-        uint64_t avg_err_fp10  = ((combined_error_fp8 >> 3) * (d_factor_fp8[subblock_idx] >> 3));
+        uint64_t avg_err_fp10  = (uint64_t)(combined_error_fp8 >> 3) * (d_factor_fp8[subblock_idx] >> 3);
         uint32_t scaled_diff16 = (uint32_t)AOMMIN(
             /*((16*avg_err)<<8)*/ (avg_err_fp10) / AOMMAX((tf_decay_factor >> 10), 1), 7 * 16);
         uint32_t adjusted_weight = (expf_tab_fp16[scaled_diff16] * TF_WEIGHT_SCALE) >> 16;
@@ -352,7 +352,7 @@ static void svt_av1_apply_temporal_filter_planewise_medium_partial_avx2(
             __m128i frame2_array      = _mm_loadl_epi64((__m128i*)(y_pre + k));
             frame2_array              = _mm_cvtepu8_epi16(frame2_array);
             __m256i frame2_array_u32  = _mm256_cvtepi16_epi32(frame2_array);
-            frame2_array_u32          = _mm256_mullo_epi32(frame2_array_u32,
+            frame2_array_u32 = _mm256_mullo_epi32(frame2_array_u32,
                                                   adjusted_weight_int32[subblock_idx_h + (j >= block_width / 2)]);
 
             accumulator_array = _mm256_add_epi32(accumulator_array, frame2_array_u32);
@@ -458,7 +458,7 @@ static void svt_av1_apply_zz_based_temporal_filter_planewise_medium_hbd_partial_
             __m256i accumulator_array = _mm256_loadu_si256((__m256i*)(y_accum + k));
             __m128i frame2_array      = _mm_loadu_si128((__m128i*)(y_pre + k));
             __m256i frame2_array_u32  = _mm256_cvtepi16_epi32(frame2_array);
-            frame2_array_u32          = _mm256_mullo_epi32(frame2_array_u32,
+            frame2_array_u32 = _mm256_mullo_epi32(frame2_array_u32,
                                                   adjusted_weight_int32[subblock_idx_h + (j >= block_width / 2)]);
 
             accumulator_array = _mm256_add_epi32(accumulator_array, frame2_array_u32);
@@ -598,7 +598,7 @@ static void svt_av1_apply_temporal_filter_planewise_medium_hbd_partial_avx2(
                                        block_error_fp8[subblock_idx]) /
             (TF_WINDOW_BLOCK_BALANCE_WEIGHT + 1);
 
-        uint64_t avg_err_fp10  = ((combined_error_fp8 >> 3) * (d_factor_fp8[subblock_idx] >> 3));
+        uint64_t avg_err_fp10  = (uint64_t)(combined_error_fp8 >> 3) * (d_factor_fp8[subblock_idx] >> 3);
         uint32_t scaled_diff16 = (uint32_t)AOMMIN(
             /*((16*avg_err)<<8)*/ (avg_err_fp10) / AOMMAX((tf_decay_factor >> 10), 1), 7 * 16);
         int adjusted_weight = (expf_tab_fp16[scaled_diff16] * TF_WEIGHT_SCALE) >> 16;
@@ -621,7 +621,7 @@ static void svt_av1_apply_temporal_filter_planewise_medium_hbd_partial_avx2(
             __m256i accumulator_array = _mm256_loadu_si256((__m256i*)(y_accum + k));
             __m128i frame2_array      = _mm_loadu_si128((__m128i*)(y_pre + k));
             __m256i frame2_array_u32  = _mm256_cvtepi16_epi32(frame2_array);
-            frame2_array_u32          = _mm256_mullo_epi32(frame2_array_u32,
+            frame2_array_u32 = _mm256_mullo_epi32(frame2_array_u32,
                                                   adjusted_weight_int32[subblock_idx_h + (j >= block_width / 2)]);
 
             accumulator_array = _mm256_add_epi32(accumulator_array, frame2_array_u32);
